+++ src/onegov.election_day/onegov/election_day/theme/election_day_theme.py
@@ -30,9 +30,7 @@ class ElectionDayTheme(BaseTheme):
             'subheader-line-height': '1.3',
             'h1-font-reduction': 'rem-calc(15)',
             'h2-font-reduction': 'rem-calc(12)',
-            'callout-panel-bg': 'scale-color({}, $lightness: 75%)'.format(
-                options['primary-color']
-            ),
+            'callout-panel-bg': '#ecfaff',
             'topbar-bg-color': '#fff',
             'topbar-dropdown-bg': '#fff',
             'topbar-link-color': '#999',