+++ src/onegov.election_day/HISTORY.rst
@@ -7,6 +7,9 @@ Changelog
 - Fixes district/entity not shown in election compounds PDF.
   [msom]
 
+- Uses a fixed callout color.
+  [msom]
+
 3.4.2 (2018-04-10)
 ~~~~~~~~~~~~~~~~~~~